+++ wasm/lib/src/module.dart
@@ -16,9 +16,10 @@ class WasmModule {
   late final Pointer<WasmerModule> _module;
 
   /// Compile a module.
-  WasmModule(Uint8List data) {
-    _module = runtime.compile(this, data);
-  }
+  WasmModule(Uint8List data) : this._(data, false);
+
+  /// Deserialize a module. See [WasmModule.serialize] for more details.
+  WasmModule.deserialize(Uint8List data) : this._(data, true);
 
   /// Returns a [WasmInstanceBuilder] that is used to add all the imports that
   /// the module needs before instantiating it.
@@ -28,6 +29,13 @@ class WasmModule {
   /// maximum number of pages.
   WasmMemory createMemory(int pages, [int? maxPages]) =>
       WasmMemory._create(pages, maxPages);
+
+  /// Serialize the module to a [Uint8List].
+  ///
+  /// This buffer can be saved to a file, sent to another device, even with a
+  /// different CPU architecture, and then passed to [WasmModule.deserialize].
+  /// Doing so speeds up start up time, by skipping compilation.
+  Uint8List serialize() => runtime.serialize(_module);
 
   /// Returns a description of all of the module's imports and exports, for
   /// debugging.
@@ -42,6 +50,10 @@ class WasmModule {
       description.write('export $exp\n');
     }
     return description.toString();
+  }
+
+  WasmModule._(Uint8List data, bool isSerialized) {
+    _module = runtime.loadModule(this, data, isSerialized);
   }
 }